Move handleHostnameError func to its own file
